+++ setup.py
@@ -1,0 +1,3 @@
+from setuptools import setup, find_packages
+
+setup(name='biblio_glutton_harvester', version='0.2', packages=find_packages())
